// src/models/StateAddFilter.ts
export const STATE_ADD_FILTER: Record<string, "RoleAndElement" | "RoleOnly"> = {
    CPN_UP_GIV_BREAK_POINT_DMG_RATIO: "RoleAndElement",
    CPN_UP_GIV_DMG_RATIO: "RoleAndElement",
    DWN_ATK_ACCUM_RATIO: "RoleAndElement",
    DWN_ATK_CONSUME_RATIO: "RoleAndElement",
    DWN_ATK_RATIO: "RoleAndElement",
    DWN_BARRIER_VALUE: "RoleAndElement",
    DWN_BREAKED_DAMAGE_RECEIVE_RATIO: "RoleAndElement",
    DWN_BREAK_DAMAGE_RECEIVE_RATIO: "RoleAndElement",
    DWN_CTD_ACCUM_RATIO: "RoleAndElement",
    DWN_CTD_CONSUME_FIXED: "RoleAndElement",
    DWN_CTD_FIXED: "RoleAndElement",
    DWN_CTD_RATIO: "RoleAndElement",
    DWN_CTR_ACCUM_RATIO: "RoleAndElement",
    DWN_CTR_CONSUME_FIXED: "RoleAndElement",
    DWN_CTR_FIXED: "RoleAndElement",
    DWN_CTR_RATIO: "RoleAndElement",
    DWN_DEF_ACCUM_RATIO: "RoleAndElement",
    DWN_DEF_RATIO: "RoleAndElement",
    DWN_ELEMENT_DMG_RATE_RATIO: "RoleOnly",
    DWN_EP_RECOVER_RATE_RATIO: "RoleAndElement",
    DWN_GIV_BLEED_DMG_RATIO: "RoleAndElement",
    DWN_GIV_BREAK_POINT_DMG_FIXED: "RoleAndElement",
    DWN_GIV_BURN_DMG_RATIO: "RoleAndElement",
    DWN_GIV_CURSE_DMG_RATIO: "RoleAndElement",
    DWN_GIV_DMG_ACCUM_RATIO: "RoleAndElement",
    DWN_GIV_DMG_CONSUME_RATIO: "RoleAndElement",
    DWN_GIV_DMG_RATIO: "RoleAndElement",
    DWN_GIV_POISON_DMG_RATIO: "RoleAndElement",
    DWN_GIV_SLIP_DMG_RATIO: "RoleAndElement",
    DWN_GIV_VORTEX_DMG_RATIO: "RoleAndElement",
    DWN_HEAL_RATE_RATIO: "RoleAndElement",
    DWN_RCV_BREAK_POINT_DMG_RATIO: "RoleAndElement",
    DWN_RCV_DMG_RATIO: "RoleAndElement",
    DWN_RCV_RECOVERY_RATIO: "RoleAndElement",
    DWN_RECOVERY_EP_RATE_RATIO: "RoleAndElement",
    DWN_SPD_ACCUM_RATIO: "RoleAndElement",
    DWN_SPD_FIXED: "RoleAndElement",
    DWN_SPD_RATIO: "RoleAndElement",
    GVE_UP_FINAL_GAIN_EP_RATIO: "RoleAndElement",
    GVE_UP_SPECIAL_ATTACK_FINAL_GIV_DMG_RATIO: "RoleAndElement",
    REGAIN_ATK: "RoleAndElement",
    REGAIN_DEF: "RoleAndElement",
    REGAIN_HP: "RoleAndElement",
    UP_ABNORMAL_HIT_RATE_RATIO: "RoleAndElement",
    UP_ABNORMAL_PARRY_RATE_RATIO: "RoleAndElement",
    UP_ATK_ACCUM_RATIO: "RoleAndElement",
    UP_ATK_CONSUME_FIXED: "RoleAndElement",
    UP_ATK_CONSUME_RATIO: "RoleAndElement",
    UP_ATK_FIXED: "RoleAndElement",
    UP_ATK_RATIO: "RoleAndElement",
    UP_BREAKED_DAMAGE_RECEIVE_RATIO: "RoleAndElement",
    UP_BREAK_DAMAGE_RECEIVE_RATIO: "RoleAndElement",
    UP_BREAK_EFFECT: "RoleAndElement",
    UP_CTD_ACCUM_RATIO: "RoleAndElement",
    UP_CTD_CONSUME_FIXED: "RoleAndElement",
    UP_CTD_FIXED: "RoleAndElement",
    UP_CTD_RATIO: "RoleAndElement",
    UP_CTR_ACCUM_RATIO: "RoleAndElement",
    UP_CTR_CONSUME_FIXED: "RoleAndElement",
    UP_CTR_FIXED: "RoleAndElement",
    UP_CTR_RATIO: "RoleAndElement",
    UP_DEF_ACCUM_RATIO: "RoleAndElement",
    UP_DEF_FIXED: "RoleAndElement",
    UP_DEF_RATIO: "RoleAndElement",
    UP_EFFECT_HIT_RATE_RATIO: "RoleAndElement",
    UP_EFFECT_PARRY_RATE_RATIO: "RoleAndElement",
    UP_ELEMENT_DMG_RATE_RATIO: "RoleOnly",
    UP_EP_RECOVER_RATE_RATIO: "RoleAndElement",
    UP_GIV_BLEED_DMG_RATIO: "RoleAndElement",
    UP_GIV_BREAK_POINT_DMG_FIXED: "RoleAndElement",
    UP_GIV_BREAK_POINT_DMG_RATIO: "RoleAndElement",
    UP_GIV_BURN_DMG_RATIO: "RoleAndElement",
    UP_GIV_CURSE_DMG_RATIO: "RoleAndElement",
    UP_GIV_DMG_ACCUM_RATIO: "RoleAndElement",
    UP_GIV_DMG_CONSUME_RATIO: "RoleAndElement",
    UP_GIV_DMG_RATIO: "RoleAndElement",
    UP_GIV_POISON_DMG_RATIO: "RoleAndElement",
    UP_GIV_SLIP_DMG_RATIO: "RoleAndElement",
    UP_GIV_VORTEX_DMG_RATIO: "RoleAndElement",
    UP_HEAL_RATE_RATIO: "RoleAndElement",
    UP_HP_FIXED: "RoleAndElement",
    UP_HP_RATIO: "RoleAndElement",
    UP_RCV_BREAK_POINT_DMG_RATIO: "RoleAndElement",
    UP_RCV_DMG_RATIO: "RoleAndElement",
    UP_RECOVERY_EP_RATE_RATIO: "RoleAndElement",
    UP_SPD_ACCUM_RATIO: "RoleAndElement",
    UP_SPD_FIXED: "RoleAndElement",
    UP_SPD_RATIO: "RoleAndElement",
    UP_WEAK_ELEMENT_DMG_ACCUM_RATIO: "RoleAndElement",
    UP_WEAK_ELEMENT_DMG_CONSUME_RATIO: "RoleAndElement",
    UP_WEAK_ELEMENT_DMG_RATIO: "RoleAndElement",
}